+++ tests/tf/aurora/test_decimation.py
@@ -6,23 +6,23 @@ Created on Thu Feb 24 14:11:24 2022
 """
 
 import unittest
-from mt_metadata.transfer_functions.fourier_coefficients import Band
+from mt_metadata.transfer_functions.processing.aurora import Decimation
 
 
-class TestBand(unittest.TestCase):
+class TestDecimation(unittest.TestCase):
     """
     Test Station metadata
     """
 
     def setUp(self):
-        self.band = Band()
+        self.decimation = Decimation()
 
     def test_initialization(self):
-        for key in self.band.get_attribute_list():
+        for key in self.decimation.get_attribute_list():
             with self.subTest(key):
                 self.assertEqual(
-                    self.band.get_attr_from_name(key),
-                    self.band._attr_dict[key]["default"],
+                    self.decimation.get_attr_from_name(key),
+                    self.decimation._attr_dict[key]["default"],
                 )
